Only create release drafts
Release mails (Watch -> Custom -> Releases) are sent out when a release is created. Unfortunately, the release notes are not yet added, so the mail is useless. 

Now the release is created in draft and after editing it, it needs to be published (one click, similar to save). With this workflow, the release email now contains the release notes.