Use component instead of directive for vue-tippy
For this, we use the <tippy> component with a content slot.
Hopefully, we'll get rid of many functions in the node composite
(useNode).

This will make the integration of content with other components easier.

The rootEl ref in DiagramRow is moved so that scroll into view continues
to work.

For the sake of readability, the light theme is now used for VueTippy.