Avoid JSON parsing the plan twice
In PlanService.fromSource() we were JSON parsing the source value to
determine if it was valid JSON and then passing the original source to
parseJson() which again parsed that value. We now only parse the value
once before calling this parseJson() method. That method is renamed as
getPlanContent(), made private and its inaccurate comment is removed.