Re-add hint for track_io_timing in Diagram
Instead of putting the hint in a tooltip in the Diagram content, it is
now displayed when hovering the disabled IO tab.

Was unexpectedly removed in commit a65a306.

Fixes #830

We can't use v-tippy directive here since tippy doesn't work on
deactivated buttons.

Extra css is required to get the wrapper span (added by the <tippy>
component) and underlying button look to be OK with btn-group.